requirements.txt
h5py==3.1.0
matplotlib==3.2.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.5.0
scikit_learn==1.1.2
scipy==1.10.1
tensorboardX==2.6
tensorflow==2.5.0
torch==1.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.60.0
